requirements.txt
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
crispy-bootstrap5==2025.4
Django==5.2
django-crispy-forms==2.4
django-polymorphic==3.1.0
django-reversion==5.1.0
django-tinymce==4.1.0
h11==0.14.0
habanero==2.2.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
libretranslatepy==2.1.1
lxml==5.4.0
mysqlclient==2.2.7
packaging==25.0
pillow==11.2.1
PyYAML==6.0.2
requests==2.32.3
setuptools==79.0.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
translate==3.6.1
tzdata==2025.2
urllib3==2.2.0
